refactor(api): add transformers.ts for mission-browse service
